generalize the pool, allow non-ByteBuffer (eg. BitBuffer) buffers
