+++ src/main/java/pbbl/BufferPool.java
@@ -31,40 +31,41 @@ import java.util.TreeMap;
 import java.util.function.Predicate;
 
 /**
- * A pool that can contain both {@code HeapByteBuffer}s and {@code DirectByteBuffer}s.
+ * A pool that can contain buffers, eg. {@link ByteBuffer}s.
  * <br><br>
- * {@link ByteBuffer}s dispatched from this pool will be reused, resulting in significant performance improvements from
- * not having to constantly allocate new {@link ByteBuffer}s.
+ * Buffers dispatched from this pool should be returned, resulting in significant performance improvements from
+ * not having to constantly allocate new buffer instances.
  *
+ * @param <T> the type of the buffers this pool stores, eg. {@link ByteBuffer}
  * @author Jacob G.
  * @since February 23, 2019
  */
-public abstract class ByteBufferPool {
+public abstract class BufferPool<T> {
     
     /**
-     * The data structure that holds all pooled {@link ByteBuffer}s.
+     * The data structure that holds all pooled buffers.
      */
-    private final NavigableMap<Integer, Deque<ByteBuffer>> buffers = new TreeMap<>();
+    private final NavigableMap<Integer, Deque<T>> buffers = new TreeMap<>();
     
     /**
      * An abstract method that, given a capacity, creates a new {@code HeapByteBuffer} or {@code DirectByteBuffer}.
      *
-     * @param n the capacity of the {@link ByteBuffer} to create.
-     * @return a newly-created {@link ByteBuffer}.
+     * @param n the capacity of the buffer to create.
+     * @return a newly-created buffer.
      */
-    protected abstract ByteBuffer create(int n);
+    protected abstract T create(int n);
     
     /**
-     * Attempts to take a new {@link ByteBuffer} from the pool.
+     * Attempts to take a new buffer from the pool.
      * <br><br>
-     * If no {@link ByteBuffer} can be found (with a capacity of at-least {@code n}) within the pool, then a new one is
+     * If no buffer can be found (with a capacity of at-least {@code n}) within the pool, then a new one is
      * created.
      *
-     * @param n the capacity of the {@link ByteBuffer} requested.
-     * @return a {@link ByteBuffer} with a capacity greater than or equal to {@code n}, with its limit set to {@code
+     * @param n the capacity of the buffer requested.
+     * @return a buffer with a capacity greater than or equal to {@code n}, with its limit set to {@code
      * n} and position set to {@code 0}.
      */
-    public ByteBuffer take(int n) {
+    public T take(int n) {
         synchronized (buffers) {
             var entry = buffers.ceilingEntry(n);
             
@@ -79,7 +80,7 @@ public abstract class ByteBufferPool {
             var deque = entry.getValue();
             
             if (!deque.isEmpty()) {
-                return deque.poll().clear().limit(n);
+                return clearAndLimitBuffer(deque.poll(), n);
             }
             
             // The first entry that was found had no ByteBuffers available, so we must now look at every greater
@@ -89,22 +90,42 @@ public abstract class ByteBufferPool {
                     .filter(Predicate.not(Deque::isEmpty))
                     .map(Deque::poll)
                     .findAny()
-                    .map(buffer -> buffer.clear().limit(n))
+                    .map(buffer -> clearAndLimitBuffer(buffer, n))
                     .orElseGet(() -> create(n));
         }
     }
     
     /**
-     * Gives the specified {@link ByteBuffer} to this {@link ByteBufferPool}.
+     * Gives the specified buffer to this {@link BufferPool}.
      * <br><br>
      * This method should only be called <strong>after</strong> calling {@link #take(int)}.
      *
-     * @param buffer the {@link ByteBuffer} to return to this pool.
+     * @param buffer the buffer to return to this pool.
      */
-    public void give(ByteBuffer buffer) {
+    public void give(T buffer) {
         synchronized (buffers) {
-            buffers.computeIfAbsent(buffer.capacity(), $ -> new ArrayDeque<>(3)).offer(buffer);
+            buffers.computeIfAbsent(bufferCapacity(buffer), $ -> new ArrayDeque<>(3)).offer(buffer);
         }
     }
+    
+    /**
+     * Clears and limits the specified buffer (in this order).
+     * 
+     * @param buffer the buffer to limit.
+     * @param limit the new limit of the buffer.
+     * @return the same {@link T} instance (the {@code buffer} parameter)
+     * @see java.nio.Buffer#clear()
+     * @see java.nio.Buffer#limit(int)
+     */
+    protected abstract T clearAndLimitBuffer(T buffer, int limit);
+    
+    /**
+     * Gets the capacity of the specified buffer.
+     *
+     * @param buffer the buffer to get the capacity of.
+     * @return the capacity of the buffer.
+     * @see java.nio.Buffer#capacity()
+     */
+    protected abstract int bufferCapacity(T buffer);
     
 }
